# optimized bbox_generators