The optimized solution calculates all the final corner coordinates upfront. It then uses a single, large torch.stack operation to arrange all these coordinate values into one flat tensor in their correct final order. A subsequent, virtually zero-cost .view() operation then reshapes this tensor into the final desired (B, 4, 2) or (B, 8, 3) format.

--- ⏱️ BENCHMARKING ---
  - Warming up...
  - Original                       0.376992 ms per run (Baseline)
  - Vectorized (single-stack)      0.109238 ms per run (3.45x faster)

--- ⏱️ BENCHMARKING ---
  - Warming up...
  - Original 3D                    0.332430 ms per run (Baseline)
  - Optimized 3D (single-stack)    0.068149 ms per run (4.88x faster)